submission group 9 - comm strategies
# Group 9 project communication strategies
We implemented different versions of MPI communication in the Fortran 2d stencil program. 

## Team members
Victoria Bauer, Sarina Danioth, Julia Dworzak

## Summary 
The purpose of this study, conducted as part of the lecture ”High Performance Computing for Weather and Climate” at ETH Zurich in the spring semester of 2022, is the implementation and performance analysis of different communication strategies using the standardised Message Passing Interface (MPI). The communication strategies were implemented into the halo update of a stencil 2D program simulating higher order diffusion. Three different experiments were conducted measuring the time needed for com- munication within the halo update of the different versions. Regarding the scalability of the problem, we found that a certain trade-off between decreasing the computational workload per worker and increasing the overall communication load in the program exists. While a parallel computing method certainly outperforms the serial computation of the stencil, we could not identify the optimal MPI communication strategy as all perform similarly well. Thus, choosing a communication strategy depends on individual needs concerning the problem and personal preferences.