Add 2D stencil with Numba
The result field is not exactly the same as the one from the original
code, but it's close enough to be able to do the benchmark analysis.
To achieve correctness we have to implement the `extend` kwarg, and for
that probably we cannot use the @stencil decorator as it is the current
version of numba.