Complete 1_lists_numpy notebook
Lists and NumPy arrays are not explained in much more detail. There are
no "weird" results that cannot be explained after learning how these two
totally objects are stored in memory. There is also a new Appendix where
we can estimate the size of the L3 cache just by checking the timings of
the copy stencil using arrays.